refactor(LadderGame): service 패키지로 분리
